requirements.txt
# Django core
Django==3.1.12
asgiref==3.11.0
pytz==2025.2
sqlparse==0.2.4
djangorestframework-simplejwt==5.2.0
django-cors-headers==4.9.0

# Image and numerical processing
Pillow==12.1.0
numpy==2.4.2
python-decouple==3.8
geopy==2.4.1
djangorestframework==3.16.1
dj-database-url==1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
python-decouple==3.8
dj-database-url-3.1.1